// ui-catalogue/src/components/molecules/Resume/types.tsx
interface Ii18nResume {
  'Resume:title': string
  'Resume:subtitle1': string
  'Resume:subtitle2': string
  'Resume:subtitle3': string
  'Resume:subtitle4': string
  'Resume:subtitle5': string
  'Resume:formajuridica': string
  'Resume:anteriorformajuridica': string
  'Resume:descripcionformajuridica': string
  'Resume:descripcionanteriorformajuridica': string
  'Resume:fechaconstitucion': string
  'Resume:fechaformajuridica': string
  'Resume:fechacambiocapitalnominal': string
  'Resume:fechadatosresumen': string
  'Resume:estadoactual': string
  'Resume:fechaestadoactual': string
  'Resume:capitalnominal': string
  'Resume:capitaldesembolsado': string
  'Resume:ventas': string
  'Resume:resultadoneto': string
  'Resume:fondospropios': string
  'Resume:capitalcirculante': string
  'Resume:activosfijos': string
  'Resume:existencias': string
  'Resume:fuentedatoseconomicos': string
  'Resume:capitalnominalanterior': string
  'Resume:numeroempleados': string
  'Resume:comentarios': string
  'Resume:nombre': string
  'Resume:domicilio': string
  'Resume:localidad': string
  'Resume:provincia': string
  'Resume:codigopostal': string
  'Resume:pais': string
  'Resume:porcentajeparticipacion': string
  'Resume:periodomediopago': string
  'Resume:periodomediocobro': string
  'Resume:deudoresdias': string
}

export interface IResumeRequestForm {
  formajuridica: string
  anteriorformajuridica: string
  descripcionformajuridica: string
  descripcionanteriorformajuridica: string
  fechaconstitucion: string
  fechaformajuridica: string
  fechacambiocapitalnominal: string
  fechadatosresumen: string
  estadoactual: string
  fechaestadoactual: string
  capitalnominal: string
  capitaldesembolsado: string
  ventas: string
  resultadoneto: string
  fondospropios: string
  capitalcirculante: string
  activosfijos: string
  existencias: string
  fuentedatoseconomicos: string
  capitalnominalanterior: string
  numeroempleados: string
  comentarios: string
  nombre1: string
  domicilio1: string
  localidad1: string
  provincia1: string
  codigopostal1: string
  pais1: string
  porcentajeparticipacion1: string
  porcentajeparticipacion2: string
  porcentajeparticipacion3: string
  porcentajeparticipacion4: string
  porcentajeparticipacion5: string
  porcentajeparticipacion6: string
  porcentajeparticipacion7: string
  porcentajeparticipacion8: string
  porcentajeparticipacion9: string
  porcentajeparticipacion10: string
  porcentajeparticipacion11: string
  porcentajeparticipacion12: string
  porcentajeparticipacion13: string
  nombre2: string
  domicilio2: string
  localidad2: string
  provincia2: string
  codigopostal2: string
  pais2: string
  nombre3: string
  domicilio3: string
  localidad3: string
  provincia3: string
  codigopostal3: string
  pais3: string
  periodomediopago: string
  periodomediocobro: string
  deudores: string
}

export interface IResume {
  i18n: Ii18nResume
  data: IResumeRequestForm
}